+++ Source/Krypton Components/Krypton.Toolkit/Controls Toolkit/KryptonNumericUpDown.cs
@@ -1922,7 +1922,11 @@ namespace Krypton.Toolkit
             if (NumericUpDown != null)
             {
                 NumericUpDown.Invalidate();
-                PI.RedrawWindow(Handle, IntPtr.Zero, IntPtr.Zero, 0x85);
+
+                if (!IsDisposed && !Disposing)
+                {
+                    PI.RedrawWindow(Handle, IntPtr.Zero, IntPtr.Zero, 0x85);
+                }
             }
         }